+++ R/provider-openai.R
@@ -315,7 +315,14 @@ method(value_turn, ProviderOpenAI) <- function(
       ContentImageInline(mime_type, output$result)
     } else if (output$type == "web_search_call") {
       # https://platform.openai.com/docs/guides/tools-web-search#output-and-citations
-      ContentToolRequestSearch(query = output$action$query, json = output)
+      first_query <- if (length(output$action$queries)) {
+        output$action$queries[[1]]
+      }
+      query <- output$action$query %||%
+        first_query %||%
+        output$action$url %||%
+        "web search"
+      ContentToolRequestSearch(query = query, json = output)
     } else {
       browser()
       cli::cli_abort(